chore(web): rename middleware to proxy (Next.js 16 convention)
